requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
greenlet==3.0.3
paho-mqtt==2.1.0
playwright==1.45.1
pyee==11.1.0
python-dotenv==1.0.1
soupsieve==2.5
typing_extensions==4.12.2
